requirements.txt
Flask==3.1.0
psycopg2==2.9.5
gunicorn==23.0.0
werkzeug==3.1.3
bcrypt==4.2.1
python-dotenv==24.3.1